refactor(fmt): Move Style closer to API representation
